web/requirements.txt
flask
flask-bcrypt
flask-socketio
Flask-SQLAlchemy
Flask-Session
python-dotenv
eventlet
redis

